+++ test/docs/frozen-date/frozen-date-setup.mjs
@@ -1,0 +1,3 @@
+import * as ck from 'chronokinesis';
+
+ck.freeze(1970, 3, 11);